Move AuthUserDTO to DTOs.User package
